ai-service/requirements.txt
# FastAPI & Server
fastapi
uvicorn
pydantic

# AI & Embeddings
groq
sentence-transformers
chromadb

# Environment
python-dotenv

# Google Gmail Integration
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# HTTP & Requests
httpx
requests

# Background Jobs
apscheduler

# PDF Parsing
pdfminer.six
pdfplumber

# Market Data & Search
yfinance
tavily-python

# Utilities (used in routes)
aiofiles